style: rename nextion_functions to display
